Backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.108.0
google-genai>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
python-dateutil>=2.8.0
pymongo>=4.6.0
dnspython>=2.4.0
